share links work
all of them are good to go. the permalink button now shows you the permalink for the article, and selects it for you by default. clicking the permalink button again will hide it. styling may need to be cleaned up, there is a height issue, but it may be a good visual queue?